server/requirements.txt
asgiref==3.3.1
blis==0.7.3
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
cymem==2.0.4
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-cors-headers==3.8.0
djangorestframework==3.12.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gunicorn==20.0.4
idna==2.10
murmurhash==1.0.4
nl-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-2.3.0/nl_core_news_sm-2.3.0.tar.gz
numpy==1.19.3
plac==1.1.3
preshed==3.0.4
psycopg2==2.8.6
pytz==2020.4
requests==2.25.0
spacy==2.3.4
sqlparse==0.4.1
srsly==1.0.4
thinc==7.4.3
tqdm==4.54.0
urllib3==1.26.2
wasabi==0.8.0
whitenoise==5.2.0
wincertstore==0.2
